@@ -85,9 +85,9 @@
     "memoize-one": "^5.1.1",
     "openseadragon": "^4.1.1",
     "pako": "^2.0.3",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-async": "^10.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-flip-toolkit": "^7.0.12",
     "react-helmet": "^6.1.0",
     "react-icons": "^4.2.0",
@@ -148,8 +148,8 @@
     "@types/node": "^20.10.6",
     "@types/openseadragon": "^3.0.10",
     "@types/pako": "^1.0.2",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.3.4",
+    "@types/react-dom": "^18.3.0",
     "@types/react-helmet": "^6.1.2",
     "@types/react-redux": "^7.1.18",
     "@types/sha1": "^1.1.3",
